TIL-22 axios api fetcher 함수 명 및 에러 출력 관련 변경하였습니다.
