+++ src/common/utils/axios/stats/toplanguageResponse.ts
@@ -1,9 +1,9 @@
 import { HttpException } from '@nestjs/common';
 import axios from 'axios';
 
-export function toplanguage_fetcher(token, variables) {
+export async function toplanguageResponse(token, variables) {
   try {
-    return axios({
+    return await axios({
       url: 'https://api.github.com/graphql',
       method: 'post',
       headers: {
@@ -37,7 +37,7 @@ export function toplanguage_fetcher(token, variables) {
   } catch (error) {
     if (error.response) {
       if (error.response.status == 401) {
-        throw new HttpException({ code: 'BAD_CREDENTIALS', message: '알수없는 토큰 입니다.' }, 401);
+        throw new HttpException({ code: 'BAD_CREDENTIALS', message: `${token} - 토큰은 만료가 되었거나, 사용할 수 없는 토큰입니다.` }, 401);
       }
       throw new HttpException({ code: 'UNKOWN', message: error.message }, 400);
     }
